Updated app.py and changed the folder structure
